+++ doc/whats_new.rst
@@ -23,7 +23,7 @@ Notable changes
 Authors
 ~~~~~~~
 
-* xxx
+* `Adam Li`_
 
 Detailed list of changes
 ~~~~~~~~~~~~~~~~~~~~~~~~
@@ -46,7 +46,7 @@ Requirements
 Bug fixes
 ^^^^^^^^^
 
-- xxx
+- Anonymization of EDF files in :func:`write_raw_bids` will now convert recording date to ``01-01-1985 00:00:00``, while setting the recording date in the ``scans.tsv`` file to an anonymized date < 1925, thus making the file EDF/EDFBrowser compliant, by `Adam Li`_ (:gh:`600`)
 
 :doc:`Find out what was new in previous releases <whats_new_previous_releases>`